Feature/ee2 final wafs (#335)
* nfcens driver script for prep was edited.

* j-jobs were moved underneath /jobs

* export LD_LIBRARY_PATH were removed.

* added SENDDBN if statement according to EE2 comment.

* Hard-coded MET version was removed.

* hardcoded path in modulefiles were removed.

* Changed "set +x" to "set -x"

* Scripts were cleaned up.

* Changed set +x to set -x

* KEEPDATA changed to "YES"

* HOMEevs now is pointing to the default path.

* emc.vpppg was replaced with $USER.

* The variables related to MET/METplus were removed.

* SENDDBN setting was modified.

* Minor edit was done for "export err".

* "exit 0" was removed.

* COMOUT path was adjusted to stop at ${evs_ver}.

* Code structure was updated (NCO requirement).

* Code structure was updated.

* Output directory structure for stats was modified.

* Output directory structure for stats was modified.

* COMINxxx changed to EVSINxxx.

* Updated COMINxxx changes.

* COMINmodel changed to COMINnfcens.

* COMINobs chnaged to COMINobsproc.

* $CYC, $cyc, $cycle, $cycles were renamed.

* inithours replaced with vhours in "stat".

* HOMEevs was modified.

* evs_ver variable was modified.

* Cat log files to stdout file.

* EE2 final: error handling, hard coded paths

* EE2 final: error handling, minor typo

* Plot legend, permission for rstprod data

* err_chk syntax was fixed.

* COMOUT set to the ptmp directory.

* References to jlogfile were removed.

* postmsg command were removed.

* Error messages were modified/postmsg was removed.

* evs_ver was fixed. cat errfile to stdout.

* Remaining EE2 items (syntax error, cpreq)

* Updating aqm driver scripts to point to the EMC AQMv7 parallel.

---------

+++ scripts/stats/wafs/exevs_wafs_atmos_stats.sh
@@ -8,7 +8,11 @@
 set -x
 
 cd $DATA
-rm wafs_stat.cmdfile
+if [ -s wafs_stat.cmdfile ];then
+	rm wafs_stat.cmdfile
+else
+	echo "WARNING: wafs_stat.cmdfile DOES NOT EXIST"
+fi
 
 export DATAsemifinal=$DATA/semifinal
 mkdir -p $DATAsemifinal
@@ -18,7 +22,7 @@ ic=0
 observations="GCIP GFS"
 for observation in $observations ; do
     if [ $observation = "GCIP" ] ; then
-	# For ICING, there are 2 different resoltions (before Nov 2023) and 3 centers
+	# For ICING, there are 2 different resolutions (before Nov 2023) and 3 centers
 	resolutions="0P25"
 	centers="blend uk us"
     elif [ $observation = "GFS" ] ; then
@@ -68,9 +72,23 @@ for resolution in $resolutions ; do
 done
 
 if [ $SENDCOM = YES ] ; then
-    mv $STATSOUTfinal/* $COMOUTfinal/.
-    # COMOUTsmall
-    mv $STATSOUTsmall/* $COMOUTsmall/.
+    cpreq $STATSOUTfinal/* $COMOUTfinal/.
+  
+    cpreq $STATSOUTsmall/* $COMOUTsmall/.
+fi
+
+#########################################
+#Cat'ing errfiles to stdout
+#########################################
+
+log_dir=$DATA/METplus_output/logs
+log_file_count=$(find $log_dir -type f |wc -l)
+if [[ $log_file_count -ne 0 ]]; then
+	for log_file in $log_dir/*; do
+		echo "Start: $log_file"
+		cat $log_file
+		echo "End: $log_file"
+	done
 fi
 #########################################
 #Cat'ing errfiles to stdout
